fix: assign UUIDs for static programs vocab too
the subjects map needs to know the ID to map stuff too for terms in the spreadsheet, some of which are these program terms, and the easiest way to add this without special handling for subject terms (e.g. enter a program_name ID in the New Value or Auth URI column even though it's neither of those) is to generate them the same way we do for other local terms, PLUS it ultimately makes sense to have UUIDs for everything rather than IDs that may become misleading over time (e.g. think of program name changes)